Truncate messages server side
