+++ CopilotKit/packages/react-core/src/openai/chat-completion-client.ts
@@ -45,59 +45,8 @@ export class ChatCompletionClient extends EventEmitter<ChatCompletionClientEvent
 
   public async fetch(params: ChatCompletionTransportFetchParams) {
     params = { ...params };
-    if (params.model && params.model in maxTokensByModel) {
-      params.maxTokens ||= maxTokensByModel[params.model];
-    } else {
-      params.maxTokens ||= DEFAULT_MAX_TOKENS;
-    }
-
     params.functions ||= [];
-    params.messages = this.buildPrompt(params);
     return await this.runPrompt(params);
-  }
-
-  private buildPrompt(params: ChatCompletionTransportFetchParams): Message[] {
-    let maxTokens = params.maxTokens!;
-    const messages = params.messages!;
-    const functions = params.functions!;
-    const functionsNumTokens = countFunctionsTokens(functions);
-    if (functionsNumTokens > maxTokens) {
-      throw new Error(`Too many tokens in function calls: ${functionsNumTokens} > ${maxTokens}`);
-    }
-    maxTokens -= functionsNumTokens;
-
-    for (const message of messages) {
-      if (message.role === "system") {
-        const numTokens = this.countTokens(message);
-        maxTokens -= numTokens;
-
-        if (maxTokens < 0) {
-          throw new Error("Not enough tokens for system message.");
-        }
-      }
-    }
-
-    const result: Message[] = [];
-    let cutoff: boolean = false;
-
-    const reversedMessages = [...messages].reverse();
-    for (const message of reversedMessages) {
-      if (message.role === "system") {
-        result.unshift(message);
-        continue;
-      } else if (cutoff) {
-        continue;
-      }
-      let numTokens = this.countTokens(message);
-      if (maxTokens < numTokens) {
-        cutoff = true;
-        continue;
-      }
-      result.unshift(message);
-      maxTokens -= numTokens;
-    }
-
-    return result;
   }
 
   private async runPrompt(params: ChatCompletionTransportFetchParams): Promise<void> {
@@ -190,40 +139,4 @@ export class ChatCompletionClient extends EventEmitter<ChatCompletionClientEvent
     this.functionCallName = "";
     this.functionCallArguments = "";
   }
-
-  public countTokens(message: Message): number {
-    if (message.content) {
-      return estimateTokens(message.content);
-    } else if (message.function_call) {
-      return estimateTokens(JSON.stringify(message.function_call));
-    }
-    return 0;
-  }
-}
-
-const maxTokensByModel: { [key: string]: number } = {
-  "gpt-3.5-turbo": 4097,
-  "gpt-3.5-turbo-16k": 16385,
-  "gpt-4": 8192,
-  "gpt-4-1106-preview": 8192,
-  "gpt-4-32k": 32768,
-  "gpt-3.5-turbo-0301": 4097,
-  "gpt-4-0314": 8192,
-  "gpt-4-32k-0314": 32768,
-  "gpt-3.5-turbo-0613": 4097,
-  "gpt-4-0613": 8192,
-  "gpt-4-32k-0613": 32768,
-  "gpt-3.5-turbo-16k-0613": 16385,
-};
-
-function estimateTokens(text: string): number {
-  return text.length / 3;
-}
-
-function countFunctionsTokens(functions: Function[]): number {
-  if (functions.length === 0) {
-    return 0;
-  }
-  const json = JSON.stringify(functions);
-  return estimateTokens(json);
 }
