+++ CopilotKit/packages/react-ui/package.json
@@ -43,8 +43,8 @@
     "typescript": "^5.1.3"
   },
   "dependencies": {
-    "@copilotkit/react-core": "workspace:*",
     "@copilotkit/shared": "workspace:*",
+    "@copilotkit/react-core": "workspace:*",
     "@heroicons/react": "^2.0.18",
     "@radix-ui/react-alert-dialog": "^1.0.4",
     "@radix-ui/react-dialog": "^1.0.4",
